enhancement: modernize the package and remove unused code, this is the base for future work
