+++ .travis.yml
@@ -6,7 +6,7 @@ env:
   - GO111MODULE=ON
 
 go:
-  - 1.12.x
+  - 1.21.x
 
 git:
   depth: 1
