+++ client_test.go
@@ -1,6 +1,7 @@
 package icapclient
 
 import (
+	"context"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -8,16 +9,16 @@ import (
 	"strconv"
 	"strings"
 	"testing"
-	"time"
 )
 
-func TestClient(t *testing.T) {
+func TestClient_Do(t *testing.T) {
 	if !testServerRunning() {
 		go startTestServer()
 	}
 
-	t.Run("Client Do RESPMOD", func(t *testing.T) {
+	t.Parallel()
 
+	t.Run("RESPMOD", func(t *testing.T) {
 		httpReq, err := http.NewRequest(http.MethodGet, "http://someurl.com", nil)
 		if err != nil {
 			t.Log(err.Error())
@@ -26,13 +27,9 @@ func TestClient(t *testing.T) {
 		}
 
 		type testSample struct {
-			httpResp            *http.Response
-			wantedStatusCode    int
-			wantedStatus        string
-			wantedTimeout       time.Duration
-			wantedDialerTimeout time.Duration
-			wantedReadTimeout   time.Duration
-			wantedWriteTimeout  time.Duration
+			httpResp         *http.Response
+			wantedStatusCode int
+			wantedStatus     string
 		}
 
 		sampleTable := []testSample{
@@ -50,12 +47,8 @@ func TestClient(t *testing.T) {
 					ContentLength: 19,
 					Body:          ioutil.NopCloser(strings.NewReader("This is a GOOD FILE")),
 				},
-				wantedStatusCode:    http.StatusNoContent,
-				wantedStatus:        "No Modifications",
-				wantedTimeout:       defaultTimeout,
-				wantedDialerTimeout: defaultTimeout,
-				wantedReadTimeout:   defaultTimeout,
-				wantedWriteTimeout:  defaultTimeout,
+				wantedStatusCode: http.StatusNoContent,
+				wantedStatus:     "No Modifications",
 			},
 			{
 				httpResp: &http.Response{
@@ -71,24 +64,20 @@ func TestClient(t *testing.T) {
 					ContentLength: 18,
 					Body:          ioutil.NopCloser(strings.NewReader("This is a BAD FILE")),
 				},
-				wantedStatusCode:    http.StatusOK,
-				wantedStatus:        "OK",
-				wantedTimeout:       defaultTimeout,
-				wantedDialerTimeout: defaultTimeout,
-				wantedReadTimeout:   defaultTimeout,
-				wantedWriteTimeout:  defaultTimeout,
+				wantedStatusCode: http.StatusOK,
+				wantedStatus:     "OK",
 			},
 		}
 
 		for _, sample := range sampleTable {
-			req, err := NewRequest(MethodRESPMOD, fmt.Sprintf("icap://localhost:%d/respmod", port), httpReq, sample.httpResp)
+			req, err := NewRequest(context.Background(), MethodRESPMOD, fmt.Sprintf("icap://localhost:%d/respmod", port), httpReq, sample.httpResp)
 			if err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
 			}
 
-			client := Client{}
+			client, _ := NewClient(Options{})
 			resp, err := client.Do(req)
 			if err != nil {
 				t.Log(err.Error())
@@ -105,57 +94,27 @@ func TestClient(t *testing.T) {
 				t.Logf("Wanted status:%s, got:%s", sample.wantedStatus, resp.Status)
 				t.Fail()
 			}
-
-			if client.Timeout != sample.wantedTimeout {
-				t.Logf("Wanted timeout to be:%v, got:%v", sample.wantedTimeout, client.Timeout)
-				t.Fail()
-			}
-			if client.scktDriver.DialerTimeout != sample.wantedDialerTimeout {
-				t.Logf("Wanted DialierTimeout to be:%v , got:%v", sample.wantedDialerTimeout, client.scktDriver.DialerTimeout)
-				t.Fail()
-			}
-			if client.scktDriver.ReadTimeout != sample.wantedReadTimeout {
-				t.Logf("Wanted ReadTimeout to be:%v , got:%v", sample.wantedReadTimeout, client.scktDriver.ReadTimeout)
-				t.Fail()
-			}
-			if client.scktDriver.WriteTimeout != sample.wantedWriteTimeout {
-				t.Logf("Wanted WriteTimeout to be:%v , got:%v", sample.wantedWriteTimeout, client.scktDriver.WriteTimeout)
-				t.Fail()
-			}
 		}
 
 	})
 
-	t.Run("Client Do REQMOD", func(t *testing.T) {
-
+	t.Run("REQMOD", func(t *testing.T) {
 		type testSample struct {
-			urlStr              string
-			wantedStatusCode    int
-			wantedStatus        string
-			wantedTimeout       time.Duration
-			wantedDialerTimeout time.Duration
-			wantedReadTimeout   time.Duration
-			wantedWriteTimeout  time.Duration
+			urlStr           string
+			wantedStatusCode int
+			wantedStatus     string
 		}
 
 		sampleTable := []testSample{
 			{
-				urlStr:              "http://goodifle.com",
-				wantedStatusCode:    http.StatusNoContent,
-				wantedStatus:        "No Modifications",
-				wantedTimeout:       defaultTimeout,
-				wantedDialerTimeout: defaultTimeout,
-				wantedReadTimeout:   defaultTimeout,
-				wantedWriteTimeout:  defaultTimeout,
+				urlStr:           "http://goodifle.com",
+				wantedStatusCode: http.StatusNoContent,
+				wantedStatus:     "No Modifications",
 			},
 			{
-				urlStr:              "http://badfile.com",
-				wantedStatusCode:    http.StatusOK,
-				wantedStatus:        "OK",
-				wantedTimeout:       defaultTimeout,
-				wantedDialerTimeout: defaultTimeout,
-				wantedReadTimeout:   defaultTimeout,
-				wantedWriteTimeout:  defaultTimeout,
+				urlStr:           "http://badfile.com",
+				wantedStatusCode: http.StatusOK,
+				wantedStatus:     "OK",
 			},
 		}
 
@@ -167,14 +126,14 @@ func TestClient(t *testing.T) {
 				return
 			}
 
-			req, err := NewRequest(MethodREQMOD, fmt.Sprintf("icap://localhost:%d/reqmod", port), httpReq, nil)
+			req, err := NewRequest(context.Background(), MethodREQMOD, fmt.Sprintf("icap://localhost:%d/reqmod", port), httpReq, nil)
 			if err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
 			}
 
-			client := Client{}
+			client, _ := NewClient(Options{})
 			resp, err := client.Do(req)
 			if err != nil {
 				t.Log(err.Error())
@@ -191,29 +150,10 @@ func TestClient(t *testing.T) {
 				t.Logf("Wanted status:%s, got:%s", sample.wantedStatus, resp.Status)
 				t.Fail()
 			}
-
-			if client.Timeout != sample.wantedTimeout {
-				t.Logf("Wanted timeout to be:%v, got:%v", sample.wantedTimeout, client.Timeout)
-				t.Fail()
-			}
-			if client.scktDriver.DialerTimeout != sample.wantedDialerTimeout {
-				t.Logf("Wanted DialierTimeout to be:%v , got:%v", sample.wantedDialerTimeout, client.scktDriver.DialerTimeout)
-				t.Fail()
-			}
-			if client.scktDriver.ReadTimeout != sample.wantedReadTimeout {
-				t.Logf("Wanted ReadTimeout to be:%v , got:%v", sample.wantedReadTimeout, client.scktDriver.ReadTimeout)
-				t.Fail()
-			}
-			if client.scktDriver.WriteTimeout != sample.wantedWriteTimeout {
-				t.Logf("Wanted WriteTimeout to be:%v , got:%v", sample.wantedWriteTimeout, client.scktDriver.WriteTimeout)
-				t.Fail()
-			}
-
 		}
 	})
 
-	t.Run("Clien Do RESPMOD with OPTIONS", func(t *testing.T) {
-
+	t.Run("RESPMOD with OPTIONS", func(t *testing.T) {
 		httpReq, err := http.NewRequest(http.MethodGet, "http://someurl.com", nil)
 		if err != nil {
 			t.Log(err.Error())
@@ -287,17 +227,16 @@ func TestClient(t *testing.T) {
 		}
 
 		for _, sample := range sampleTable {
-
 			urlStr := fmt.Sprintf("icap://localhost:%d/respmod", port)
 
-			optReq, err := NewRequest(MethodOPTIONS, urlStr, nil, nil)
+			optReq, err := NewRequest(context.Background(), MethodOPTIONS, urlStr, nil, nil)
 			if err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
 			}
 
-			client := Client{}
+			client, _ := NewClient(Options{})
 			optResp, err := client.Do(optReq)
 			if err != nil {
 				t.Log(err.Error())
@@ -334,14 +273,14 @@ func TestClient(t *testing.T) {
 
 			}
 
-			req, err := NewRequest(MethodRESPMOD, urlStr, httpReq, sample.httpResp)
+			req, err := NewRequest(context.Background(), MethodRESPMOD, urlStr, httpReq, sample.httpResp)
 			if err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
 			}
 
-			if err := req.ExtendHeader(optResp.Header); err != nil {
+			if err := req.extendHeader(optResp.Header); err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
@@ -367,7 +306,7 @@ func TestClient(t *testing.T) {
 		}
 	})
 
-	t.Run("Client Do REQMOD with OPTIONS", func(t *testing.T) {
+	t.Run("REQMOD with OPTIONS", func(t *testing.T) {
 		type testSample struct {
 			urlStr                 string
 			wantedStatusCode       int
@@ -408,16 +347,16 @@ func TestClient(t *testing.T) {
 
 		for _, sample := range sampleTable {
 
-			client := Client{}
 			urlStr := fmt.Sprintf("icap://localhost:%d/reqmod", port)
 
-			optReq, err := NewRequest(MethodOPTIONS, urlStr, nil, nil)
+			optReq, err := NewRequest(context.Background(), MethodOPTIONS, urlStr, nil, nil)
 			if err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
 			}
 
+			client, _ := NewClient(Options{})
 			optResp, err := client.Do(optReq)
 			if err != nil {
 				t.Log(err.Error())
@@ -453,14 +392,14 @@ func TestClient(t *testing.T) {
 				return
 			}
 
-			req, err := NewRequest(MethodREQMOD, urlStr, httpReq, nil)
+			req, err := NewRequest(context.Background(), MethodREQMOD, urlStr, httpReq, nil)
 			if err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
 			}
 
-			if err := req.ExtendHeader(optResp.Header); err != nil {
+			if err := req.extendHeader(optResp.Header); err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
@@ -489,33 +428,21 @@ func TestClient(t *testing.T) {
 	t.Run("Client Do REQMOD with Custom Driver", func(t *testing.T) {
 
 		type testSample struct {
-			urlStr              string
-			wantedStatusCode    int
-			wantedStatus        string
-			wantedTimeout       time.Duration
-			wantedDialerTimeout time.Duration
-			wantedReadTimeout   time.Duration
-			wantedWriteTimeout  time.Duration
+			urlStr           string
+			wantedStatusCode int
+			wantedStatus     string
 		}
 
 		sampleTable := []testSample{
 			{
-				urlStr:              "http://goodifle.com",
-				wantedStatusCode:    http.StatusNoContent,
-				wantedStatus:        "No Modifications",
-				wantedTimeout:       defaultTimeout,
-				wantedDialerTimeout: 2 * time.Second,
-				wantedReadTimeout:   2 * time.Second,
-				wantedWriteTimeout:  2 * time.Second,
+				urlStr:           "http://goodifle.com",
+				wantedStatusCode: http.StatusNoContent,
+				wantedStatus:     "No Modifications",
 			},
 			{
-				urlStr:              "http://badfile.com",
-				wantedStatusCode:    http.StatusOK,
-				wantedStatus:        "OK",
-				wantedTimeout:       defaultTimeout,
-				wantedDialerTimeout: 2 * time.Second,
-				wantedReadTimeout:   2 * time.Second,
-				wantedWriteTimeout:  2 * time.Second,
+				urlStr:           "http://badfile.com",
+				wantedStatusCode: http.StatusOK,
+				wantedStatus:     "OK",
 			},
 		}
 
@@ -527,21 +454,14 @@ func TestClient(t *testing.T) {
 				return
 			}
 
-			req, err := NewRequest(MethodREQMOD, fmt.Sprintf("icap://localhost:%d/reqmod", port), httpReq, nil)
+			req, err := NewRequest(context.Background(), MethodREQMOD, fmt.Sprintf("icap://localhost:%d/reqmod", port), httpReq, nil)
 			if err != nil {
 				t.Log(err.Error())
 				t.Fail()
 				return
 			}
 
-			client := Client{}
-			client.SetDriver(&Driver{
-				Host:          "127.0.0.1",
-				Port:          1344,
-				DialerTimeout: 2 * time.Second,
-				ReadTimeout:   2 * time.Second,
-				WriteTimeout:  2 * time.Second,
-			})
+			client, _ := NewClient(Options{})
 			resp, err := client.Do(req)
 			if err != nil {
 				t.Log(err.Error())
@@ -556,23 +476,6 @@ func TestClient(t *testing.T) {
 
 			if resp.Status != sample.wantedStatus {
 				t.Logf("Wanted status:%s, got:%s", sample.wantedStatus, resp.Status)
-				t.Fail()
-			}
-
-			if client.Timeout != sample.wantedTimeout {
-				t.Logf("Wanted timeout to be:%v, got:%v", sample.wantedTimeout, client.Timeout)
-				t.Fail()
-			}
-			if client.scktDriver.DialerTimeout != sample.wantedDialerTimeout {
-				t.Logf("Wanted DialierTimeout to be:%v , got:%v", sample.wantedDialerTimeout, client.scktDriver.DialerTimeout)
-				t.Fail()
-			}
-			if client.scktDriver.ReadTimeout != sample.wantedReadTimeout {
-				t.Logf("Wanted ReadTimeout to be:%v , got:%v", sample.wantedReadTimeout, client.scktDriver.ReadTimeout)
-				t.Fail()
-			}
-			if client.scktDriver.WriteTimeout != sample.wantedWriteTimeout {
-				t.Logf("Wanted WriteTimeout to be:%v , got:%v", sample.wantedWriteTimeout, client.scktDriver.WriteTimeout)
 				t.Fail()
 			}
