+++ examples/respmod.go
@@ -1,6 +1,7 @@
 package examples
 
 import (
+	"context"
 	"fmt"
 	"log"
 	"net/http"
@@ -10,60 +11,60 @@ import (
 )
 
 func makeRespmodCall() {
+	ctx := context.Background()
 
-	/* preparing the http request required for the RESPMOD */
+	// prepare the http request
 	httpReq, err := http.NewRequest(http.MethodGet, "http://localhost:8000/sample.pdf", nil)
-
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	/* making the http client & making the request call to get the response needed for the icap RESPMOD call */
+	// create a new http client
 	httpClient := &http.Client{}
 
+	// send the http request
 	httpResp, err := httpClient.Do(httpReq)
-
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	/* making a icap request with OPTIONS method */
-	optReq, err := ic.NewRequest(ic.MethodOPTIONS, "icap://127.0.0.1:1344/respmod", nil, nil)
-
+	// create a new icap OPTIONS request
+	optReq, err := ic.NewRequest(ctx, ic.MethodOPTIONS, "icap://127.0.0.1:1344/respmod", nil, nil)
 	if err != nil {
 		log.Fatal(err)
-		return
 	}
 
-	/* making the icap client responsible for making the requests */
-	client := &ic.Client{
+	// create the icap client
+	client, err := ic.NewClient(ic.Options{
 		Timeout: 5 * time.Second,
+	})
+	if err != nil {
+		log.Fatal(err)
 	}
 
-	/* making the OPTIONS request call */
+	// send the OPTIONS request
 	optResp, err := client.Do(optReq)
-
-	if err != nil {
-		log.Fatal(err)
-		return
-	}
-
-	/* making a icap request with RESPMOD method */
-	req, err := ic.NewRequest(ic.MethodRESPMOD, "icap://127.0.0.1:1344/respmod", httpReq, httpResp)
-
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	req.SetPreview(optResp.PreviewBytes) // setting the preview bytes obtained from the OPTIONS call
+	// create a new icap RESPMOD request
+	req, err := ic.NewRequest(ctx, ic.MethodRESPMOD, "icap://127.0.0.1:1344/respmod", httpReq, httpResp)
+	if err != nil {
+		log.Fatal(err)
+	}
 
-	/* making the RESPMOD request call */
+	// set the preview bytes obtained from the OPTIONS call
+	err = req.SetPreview(optResp.PreviewBytes)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// send the RESPMOD request
 	resp, err := client.Do(req)
-
 	if err != nil {
 		log.Fatal(err)
 	}
 
 	fmt.Println(resp.StatusCode)
-
 }
