+++ go.mod
@@ -1,8 +1,5 @@
 module github.com/egirna/icap-client
 
-go 1.12
+go 1.21
 
-require (
-	github.com/davecgh/go-spew v1.1.1
-	github.com/egirna/icap v0.0.0-20181108071049-d5ee18bd70bc
-)
+require github.com/egirna/icap v0.0.0-20181108071049-d5ee18bd70bc
