--- methods.go
@@ -1,9 +1,0 @@
-package icapclient
-
-var (
-	registeredMethods = map[string]bool{
-		MethodOPTIONS: true,
-		MethodRESPMOD: true,
-		MethodREQMOD:  true,
-	}
-)
