+++ request.go
@@ -1,11 +1,16 @@
 package icapclient
 
 import (
+	"bytes"
 	"context"
-	"fmt"
+	"errors"
+	"io"
+	"io/ioutil"
 	"net/http"
-	"net/http/httputil"
 	"net/url"
+	"os"
+	"slices"
+	"strconv"
 	"strings"
 )
 
@@ -18,14 +23,14 @@ type Request struct {
 	HTTPResponse          *http.Response
 	ChunkLength           int
 	PreviewBytes          int
-	ctx                   *context.Context
+	ctx                   context.Context
 	previewSet            bool
 	bodyFittedInPreview   bool
 	remainingPreviewBytes []byte
 }
 
-// NewRequest is the factory function for Request
-func NewRequest(method, urlStr string, httpReq *http.Request, httpResp *http.Response) (*Request, error) {
+// NewRequest returns a new Request given a context, method, url, http request and http response
+func NewRequest(ctx context.Context, method, urlStr string, httpReq *http.Request, httpResp *http.Response) (*Request, error) {
 
 	method = strings.ToUpper(method)
 
@@ -41,120 +46,175 @@ func NewRequest(method, urlStr string, httpReq *http.Request, httpResp *http.Res
 		Header:       make(map[string][]string),
 		HTTPRequest:  httpReq,
 		HTTPResponse: httpResp,
+		ctx:          ctx,
 	}
 
-	if err := req.Validate(); err != nil {
+	if err := req.validate(); err != nil {
 		return nil, err
 	}
 
 	return req, nil
 }
 
-// DumpRequest returns the given request in its ICAP/1.x wire
-// representation.
-func DumpRequest(req *Request) ([]byte, error) {
+// SetPreview sets the preview bytes in the icap header
+func (r *Request) SetPreview(maxBytes int) error {
 
-	// Making the ICAP message block
+	bodyBytes := []byte{}
 
-	reqStr := fmt.Sprintf("%s %s %s%s", req.Method, req.URL.String(), ICAPVersion, CRLF)
+	previewBytes := 0
 
-	for headerName, vals := range req.Header {
-		for _, val := range vals {
-			reqStr += fmt.Sprintf("%s: %s%s", headerName, val, CRLF)
+	// receiving the body bites to determine the preview bytes depending on the request ICAP method
+	if r.Method == MethodREQMOD {
+		if r.HTTPRequest == nil {
+			return nil
 		}
-	}
+		if r.HTTPRequest.Body != nil {
+			var err error
+			bodyBytes, err = ioutil.ReadAll(r.HTTPRequest.Body)
 
-	reqStr += "Encapsulated: %s" + CRLF // will populate the Encapsulated header value after making the http Request & Response messages
-	reqStr += CRLF
-
-	// Making the HTTP Request message block
-
-	httpReqStr := ""
-	if req.HTTPRequest != nil {
-		b, err := httputil.DumpRequestOut(req.HTTPRequest, true)
-
-		if err != nil {
-			return nil, err
-		}
-
-		httpReqStr += string(b)
-		replaceRequestURIWithActualURL(&httpReqStr, req.HTTPRequest.URL.EscapedPath(), req.HTTPRequest.URL.String())
-
-		if req.Method == MethodREQMOD {
-			if req.previewSet {
-				parsePreviewBodyBytes(&httpReqStr, req.PreviewBytes)
+			if err != nil {
+				return err
 			}
 
-			if !bodyAlreadyChunked(httpReqStr) {
-				headerStr, bodyStr, ok := splitBodyAndHeader(httpReqStr)
-				if ok {
-					addHexaBodyByteNotations(&bodyStr)
-					mergeHeaderAndBody(&httpReqStr, headerStr, bodyStr)
-				}
+			defer r.HTTPRequest.Body.Close()
+		}
+	}
+
+	if r.Method == MethodRESPMOD {
+		if r.HTTPResponse == nil {
+			return nil
+		}
+
+		if r.HTTPResponse.Body != nil {
+			var err error
+			bodyBytes, err = ioutil.ReadAll(r.HTTPResponse.Body)
+
+			if err != nil {
+				return err
 			}
 
+			defer r.HTTPResponse.Body.Close()
 		}
-
-		if httpReqStr != "" { // if the HTTP Request message block doesn't end with a \r\n\r\n, then going to add one by force for better calculation of byte offsets
-			for !strings.HasSuffix(httpReqStr, DoubleCRLF) {
-				httpReqStr += CRLF
-			}
-		}
-
 	}
 
-	// Making the HTTP Response message block
+	previewBytes = len(bodyBytes)
 
-	httpRespStr := ""
-	if req.HTTPResponse != nil {
-		b, err := httputil.DumpResponse(req.HTTPResponse, true)
-
-		if err != nil {
-			return nil, err
-		}
-
-		httpRespStr += string(b)
-
-		if req.previewSet {
-			parsePreviewBodyBytes(&httpRespStr, req.PreviewBytes)
-		}
-
-		if !bodyAlreadyChunked(httpRespStr) {
-			headerStr, bodyStr, ok := splitBodyAndHeader(httpRespStr)
-			if ok {
-				addHexaBodyByteNotations(&bodyStr)
-				mergeHeaderAndBody(&httpRespStr, headerStr, bodyStr)
-			}
-		}
-
-		if httpRespStr != "" && !strings.HasSuffix(httpRespStr, DoubleCRLF) { // if the HTTP Response message block doesn't end with a \r\n\r\n, then going to add one by force for better calculation of byte offsets
-			httpRespStr += CRLF
-		}
-
+	// if the preview byte is 0 or less, there is no question of the body-fitting insides
+	if previewBytes > 0 {
+		r.bodyFittedInPreview = true
 	}
 
-	if encpVal := req.Header.Get(EncapsulatedHeader); encpVal != "" {
-		reqStr = fmt.Sprintf(reqStr, encpVal)
-	} else {
-		//populating the Encapsulated header of the ICAP message portion
-		setEncapsulatedHeaderValue(&reqStr, httpReqStr, httpRespStr)
+	// if the preview bytes are greater than what was mentioned by the ICAP Server (did not fit in the body)
+	if previewBytes > maxBytes {
+		previewBytes = maxBytes
+		r.bodyFittedInPreview = false
+		// storing the rest of the body byte which was not sent as preview for further operations
+		r.remainingPreviewBytes = bodyBytes[maxBytes:]
 	}
 
-	// determining if the http message needs the full body fitted in the preview portion indicator or not
-	if httpRespStr != "" && req.previewSet && req.bodyFittedInPreview {
-		addFullBodyInPreviewIndicator(&httpRespStr)
+	// set the body to the http message depending on the request method
+	if r.Method == MethodREQMOD {
+		r.HTTPRequest.Body = io.NopCloser(bytes.NewReader(bodyBytes))
 	}
 
-	if req.Method == MethodREQMOD && req.previewSet && req.bodyFittedInPreview {
-		addFullBodyInPreviewIndicator(&httpReqStr)
+	if r.Method == MethodRESPMOD {
+		r.HTTPResponse.Body = io.NopCloser(bytes.NewReader(bodyBytes))
 	}
 
-	data := []byte(reqStr + httpReqStr + httpRespStr)
+	// assign preview byte information to the header
+	r.Header.Set("Preview", strconv.Itoa(previewBytes))
+	r.PreviewBytes = previewBytes
+	r.previewSet = true
 
-	return data, nil
+	return nil
 }
 
-// SetContext sets a context for the ICAP request
-func (r *Request) SetContext(ctx context.Context) { // TODO: make context take control over the whole operation
-	r.ctx = &ctx
+// setDefaultRequestHeaders is called by the client before sending the request
+// to the ICAP server to ensure all required headers are set
+func (r *Request) setDefaultRequestHeaders() {
+	if _, exists := r.Header["Allow"]; !exists {
+		r.Header.Add("Allow", "204") // assigning 204 by default if Allow not provided
+	}
+
+	if _, exists := r.Header["Host"]; !exists {
+		hostName, _ := os.Hostname()
+		r.Header.Add("Host", hostName)
+	}
+}
+
+// extendHeader extends the current ICAP Request header with a new header
+func (r *Request) extendHeader(hdr http.Header) error {
+	for header, values := range hdr {
+
+		if header == previewHeader && r.previewSet {
+			continue
+		}
+
+		if header == encapsulatedHeader {
+			continue
+		}
+
+		for _, value := range values {
+			if header == previewHeader {
+				pb, err := strconv.Atoi(value)
+				if err != nil {
+					return err
+				}
+				if err := r.SetPreview(pb); err != nil {
+					return err
+				}
+				continue
+			}
+			r.Header.Add(header, value)
+		}
+	}
+
+	return nil
+}
+
+// validate checks if the ICAP request is valid or not
+func (r *Request) validate() error {
+	var err error
+
+	// check if the ICAP request has a context
+	if r.ctx == nil {
+		err = errors.Join(err, ErrNoContext)
+	}
+
+	// check if the ICAP request method is allowed
+	if methodAllowed := slices.Contains([]string{
+		MethodOPTIONS,
+		MethodRESPMOD,
+		MethodREQMOD,
+	}, r.Method); !methodAllowed {
+		err = errors.Join(err, ErrMethodNotAllowed)
+	}
+
+	// check if the ICAP url is valid and contains all required fields
+	{
+		if r.URL.Scheme != schemeICAP {
+			err = errors.Join(err, ErrInvalidScheme)
+		}
+
+		if r.URL.Host == "" {
+			err = errors.Join(err, ErrInvalidHost)
+		}
+	}
+
+	// check if the ICAP request method is aligned with the http messages
+	{
+		if r.Method == MethodREQMOD && r.HTTPRequest == nil {
+			err = errors.Join(err, ErrREQMODWithoutReq)
+		}
+
+		if r.Method == MethodREQMOD && r.HTTPResponse != nil {
+			err = errors.Join(err, ErrREQMODWithResp)
+		}
+
+		if r.Method == MethodRESPMOD && r.HTTPResponse == nil {
+			err = errors.Join(err, ErrRESPMODWithoutResp)
+		}
+	}
+
+	return err
 }
