+++ response.go
@@ -2,7 +2,7 @@ package icapclient
 
 import (
 	"bufio"
-	"errors"
+	"fmt"
 	"net/http"
 	"strconv"
 	"strings"
@@ -18,25 +18,8 @@ type Response struct {
 	ContentResponse *http.Response
 }
 
-var (
-	optionValues = map[string]bool{
-		PreviewHeader:          true,
-		MethodsHeader:          true,
-		AllowHeader:            true,
-		TransferPreviewHeader:  true,
-		ServiceHeader:          true,
-		ISTagHeader:            true,
-		OptBodyTypeHeader:      true,
-		MaxConnectionsHeader:   true,
-		OptionsTTLHeader:       true,
-		ServiceIDHeader:        true,
-		TransferIgnoreHeader:   true,
-		TransferCompleteHeader: true,
-	}
-)
-
-// ReadResponse converts a Reader to a icapclient Response
-func ReadResponse(b *bufio.Reader) (*Response, error) {
+// readResponse reads the response from the icap server
+func readResponse(b *bufio.Reader) (*Response, error) {
 
 	resp := &Response{
 		Header: make(map[string][]string),
@@ -46,17 +29,18 @@ func ReadResponse(b *bufio.Reader) (*Response, error) {
 	httpMsg := ""
 	for currentMsg, err := b.ReadString('\n'); err == nil || currentMsg != ""; currentMsg, err = b.ReadString('\n') { // keep reading the buffer message which is the http response message
 
-		if isRequestLine(currentMsg) { // if the current message line if the first line of the message portion(request line)
+		// if the current message line if the first line of the message portion(request line)
+		if isRequestLine(currentMsg) {
 			ss := strings.Split(currentMsg, " ")
 
-			if len(ss) < 3 { // must contain 3 words, for example: "ICAP/1.0 200 OK" or "GET /something HTTP/1.1"
-				return nil, errors.New(ErrInvalidTCPMsg + ":" + currentMsg)
+			// must contain 3 words, for example, "ICAP/1.0 200 OK" or "GET /something HTTP/1.1"
+			if len(ss) < 3 {
+				return nil, fmt.Errorf("%w: %s", ErrInvalidTCPMsg, currentMsg)
 			}
 
 			// preparing the scheme below
-
-			if ss[0] == ICAPVersion {
-				scheme = SchemeICAP
+			if ss[0] == icapVersion {
+				scheme = schemeICAP
 				resp.StatusCode, resp.Status, err = getStatusWithCode(ss[1], strings.Join(ss[2:], " "))
 				if err != nil {
 					return nil, err
@@ -64,35 +48,41 @@ func ReadResponse(b *bufio.Reader) (*Response, error) {
 				continue
 			}
 
-			if ss[0] == HTTPVersion {
-				scheme = SchemeHTTPResp
+			if ss[0] == httpVersion {
+				scheme = schemeHTTPResp
 				httpMsg = ""
 			}
 
-			if strings.TrimSpace(ss[2]) == HTTPVersion { // for a http request message if the scheme version is always at last, for example: GET /something HTTP/1.1
-				scheme = SchemeHTTPReq
+			// http request message scheme version should always be at the end,
+			// for example, GET /something HTTP/1.1
+			if strings.TrimSpace(ss[2]) == httpVersion {
+				scheme = schemeHTTPReq
 				httpMsg = ""
 			}
 		}
 
-		// preparing the header for ICAP & contents for HTTP messages below
-
-		if scheme == SchemeICAP {
-			if currentMsg == LF || currentMsg == CRLF { // don't want to count the Line Feed as header
+		// preparing the header for ICAP & contents for the HTTP messages below
+		if scheme == schemeICAP {
+			// ignore the CRLF and the LF, shouldn't be counted
+			if currentMsg == lf || currentMsg == crlf {
 				continue
 			}
+
 			header, val := getHeaderVal(currentMsg)
-			if header == PreviewHeader {
+			if header == previewHeader {
 				pb, _ := strconv.Atoi(val)
 				resp.PreviewBytes = pb
 			}
+
 			resp.Header.Add(header, val)
 		}
 
-		if scheme == SchemeHTTPReq {
-			httpMsg += strings.TrimSpace(currentMsg) + CRLF
+		if scheme == schemeHTTPReq {
+			httpMsg += strings.TrimSpace(currentMsg) + crlf
 			bufferEmpty := b.Buffered() == 0
-			if currentMsg == CRLF || bufferEmpty { // a CRLF indicates the end of a http message and the buffer check is just in case the buffer eneded with one last message instead of a CRLF
+
+			// a crlf indicates the end of the HTTP message and the buffer check is just in case the buffer ended with one last message instead of a crlf
+			if currentMsg == crlf || bufferEmpty {
 				var erR error
 				resp.ContentRequest, erR = http.ReadRequest(bufio.NewReader(strings.NewReader(httpMsg)))
 				if erR != nil {
@@ -102,10 +92,10 @@ func ReadResponse(b *bufio.Reader) (*Response, error) {
 			}
 		}
 
-		if scheme == SchemeHTTPResp {
-			httpMsg += strings.TrimSpace(currentMsg) + CRLF
+		if scheme == schemeHTTPResp {
+			httpMsg += strings.TrimSpace(currentMsg) + crlf
 			bufferEmpty := b.Buffered() == 0
-			if currentMsg == CRLF || bufferEmpty {
+			if currentMsg == crlf || bufferEmpty {
 				var erR error
 				resp.ContentResponse, erR = http.ReadResponse(bufio.NewReader(strings.NewReader(httpMsg)), resp.ContentRequest)
 				if erR != nil {
@@ -119,5 +109,4 @@ func ReadResponse(b *bufio.Reader) (*Response, error) {
 	}
 
 	return resp, nil
-
 }
